+++ src/executor/mod.rs
@@ -1,5 +1,11 @@
-//! Load test executor 
+//! Load test executor
 
 pub mod limit;
-mod future;
+mod measured_future;
 
+use std::future::Future;
+use std::io::Result;
+
+use crate::report::Event;
+
+pub use measured_future::MeasuredFuture;