+++ src/report/event.rs
@@ -1,5 +1,7 @@
-use super::{EventProcessor, Event, EventType};
-use crate::time::{Duration, Instant, cmp_instant_with_delta};
+use std::io::ErrorKind;
+
+use super::{Event, EventProcessor, EventType};
+use crate::time::{cmp_instant_with_delta, Duration, Instant};
 
 static EVENT_DELTA: Duration = Duration::from_millis(2);
 
@@ -19,22 +21,26 @@ impl Event {
     }
 
     /// Creates a success event
-    /// 
+    ///
     /// ```rust
     /// use profusion::{report::Event, report::EventType, time::Instant};
-    /// 
+    ///
     /// let event = Event::success("default", Instant::now(), Instant::now());
     /// assert_eq!(event.kind(), EventType::Success);
     /// ```
-    pub fn success(name: &'static str, started_at: Instant, finished_at: Instant) -> Self {
+    pub fn success(
+        name: &'static str,
+        started_at: Instant,
+        finished_at: Instant,
+    ) -> Self {
         Self::new(name, started_at, finished_at, EventType::Success)
     }
 
     /// Creates an error event
-    /// 
+    ///
     /// ```rust
     /// use profusion::{report::Event, report::EventType, time::Instant};
-    /// 
+    ///
     /// let (start, end) = (Instant::now(), Instant::now());
     /// let event = Event::error("default", start, end);
     /// assert_eq!(event.kind(), EventType::Error);
@@ -44,14 +50,18 @@ impl Event {
     }
 
     /// Creates a timeout event
-    /// 
+    ///
     /// ```rust
     /// use profusion::{report::Event, report::EventType, time::Instant};
-    /// 
+    ///
     /// let event = Event::error("default", Instant::now(), Instant::now());
     /// assert_eq!(event.kind(), EventType::Error);
     /// ```
-    pub fn timeout(name: &'static str, started_at: Instant, finished_at: Instant) -> Self {
+    pub fn timeout(
+        name: &'static str,
+        started_at: Instant,
+        finished_at: Instant,
+    ) -> Self {
         Self::new(name, started_at, finished_at, EventType::Timeout)
     }
 
@@ -68,14 +78,14 @@ impl Event {
             }
         }
     }
-    
+
     /// Calculates latency based on event time span
-    /// 
+    ///
     /// ```rust
     /// use profusion::{report::Event, report::EventType, time::Instant, time::Duration};
-    /// 
+    ///
     /// let start = Instant::now();
-    /// let end = start + Duration::from_secs(1); 
+    /// let end = start + Duration::from_secs(1);
     //
     /// let event = Event::error("default", start, end);
     /// assert_eq!(event.latency(), Duration::from_secs(1));
@@ -85,7 +95,7 @@ impl Event {
     }
 
     /// Type of the event that was captured
-    /// 
+    ///
     /// ```rust
     /// use profusion::{report::Event, report::EventType, time::Instant};
     ///
@@ -97,7 +107,7 @@ impl Event {
     }
 
     /// Returns event name
-    /// 
+    ///
     /// ```rust
     /// use profusion::{Event, time::Instant};
     ///
@@ -111,10 +121,10 @@ impl Event {
 }
 
 /// Creates successfull event from tuple of name and two `Instant` objects
-/// 
+///
 /// ```rust
 /// use profusion::{report::Event, report::EventType, time::Instant};
-/// 
+///
 /// let event = Event::from(("custom_event_name", Instant::now(), Instant::now()));
 /// assert_eq!(event.kind(), EventType::Success)
 /// ```
@@ -125,10 +135,10 @@ impl From<(&'static str, Instant, Instant)> for Event {
 }
 
 /// Creates successfull event from tuple of name, start time and duration
-/// 
+///
 /// ```rust
 /// use profusion::{report::Event, time::Instant, time::Duration};
-/// 
+///
 /// let event = Event::from(("custom_event_name", Instant::now(), Duration::from_millis(100)));
 /// assert_eq!(event.latency(), Duration::from_millis(100))
 /// ```
@@ -138,23 +148,33 @@ impl From<(&'static str, Instant, Duration)> for Event {
     }
 }
 
-/// Creates event based on IO error
-/// 
+/// Creates event based on IO result
+///
 /// ```rust
 /// use profusion::{report::Event, report::EventType, time::Instant};
-/// use std::io::{Error, ErrorKind};
-/// 
-/// let event_timeout = Event::from(("timeout_event", Instant::now(), Instant::now(), &Error::from(ErrorKind::TimedOut)));
-/// let event_error = Event::from(("error_event", Instant::now(), Instant::now(), &Error::from(ErrorKind::AddrInUse)));
+/// use std::io::{Result, Error, ErrorKind};
 ///
-/// assert_eq!(event_timeout.kind(), EventType::Timeout);
-/// assert_eq!(event_error.kind(), EventType::Error);
+/// let events: Vec<Event> = vec![
+///    ("timeout_event", Instant::now(), Instant::now(), &Result::<u32>::Err(Error::from(ErrorKind::TimedOut))).into(),
+///    ("error_event", Instant::now(), Instant::now(), &Result::<u32>::Err(Error::from(ErrorKind::AddrInUse))).into(),
+///    ("success_event", Instant::now(), Instant::now(), &Result::<u32>::Ok(123)).into()
+/// ];
+///
+/// assert_eq!(
+///     events,
+///     vec![
+///         Event::timeout("timeout_event", Instant::now(), Instant::now()),    
+///         Event::error("error_event", Instant::now(), Instant::now()),
+///         Event::success("success_event", Instant::now(), Instant::now()),
+///     ]
+/// );
 /// ```
-impl From<(&'static str, Instant, Instant, &std::io::Error)> for Event {
-    fn from(value: (&'static str, Instant, Instant, &std::io::Error)) -> Self {
-        let kind = match value.3.kind() {
-            std::io::ErrorKind::TimedOut => EventType::Timeout,
-            _ => EventType::Error
+impl<T> From<(&'static str, Instant, Instant, &std::io::Result<T>)> for Event {
+    fn from(value: (&'static str, Instant, Instant, &std::io::Result<T>)) -> Self {
+        let kind = match value.3 {
+            Ok(..) => EventType::Success,
+            Err(error) if error.kind() == ErrorKind::TimedOut => EventType::Timeout,
+            Err(..) => EventType::Error,
         };
 
         Self::new(value.0, value.1, value.2, kind)
@@ -162,7 +182,7 @@ impl From<(&'static str, Instant, Instant, &std::io::Error)> for Event {
 }
 
 /// Equality implementation for Event with time delta of 1ms
-/// 
+///
 /// Events are going to be equal when start or end time is within 1ms of each other
 ///
 /// # Example
@@ -171,7 +191,7 @@ impl From<(&'static str, Instant, Instant, &std::io::Error)> for Event {
 ///
 /// let first_time = Instant::now();
 /// let first_with_below_ms = first_time + Duration::from_micros(999);
-/// 
+///
 /// assert_eq!(
 ///    Event::from(("delta_time", first_time, Duration::from_millis(100))),
 ///    Event::from(("delta_time", first_with_below_ms, Duration::from_millis(100))),
@@ -186,9 +206,7 @@ impl PartialEq for Event {
     }
 }
 
-impl Eq for Event {
-
-}
+impl Eq for Event {}
 
 #[cfg(test)]
 mod tests {
@@ -201,7 +219,7 @@ mod tests {
     fn create_time_pair(duration: Option<Duration>) -> (Instant, Instant) {
         let duration = match duration {
             Some(duration) => duration,
-            None => Duration::from_secs(1)
+            None => Duration::from_secs(1),
         };
 
         let time = Instant::now();
@@ -213,12 +231,7 @@ mod tests {
     fn calculates_latency_from_instant_difference() {
         let (start, end) = create_time_pair(Some(Duration::from_millis(100)));
 
-        let event = Event::new(
-            "something",
-            start,
-            end,
-            EventType::Success,
-        );
+        let event = Event::new("something", start, end, EventType::Success);
 
         assert_eq!(event.latency(), Duration::from_millis(100));
     }
@@ -279,7 +292,12 @@ mod tests {
 
         assert_eq!(
             Event::new("event1", start + EVENT_DELTA, end, EventType::Success),
-            Event::new("event1", start, end + EVENT_DELTA - NANOSECOND, EventType::Success)
+            Event::new(
+                "event1",
+                start,
+                end + EVENT_DELTA - NANOSECOND,
+                EventType::Success
+            )
         );
     }
 
@@ -301,7 +319,7 @@ mod tests {
             Event::success(
                 "event3",
                 time.clone() + Duration::from_millis(10),
-                time.clone() + Duration::from_millis(30)
+                time.clone() + Duration::from_millis(30),
             ),
             Event::timeout(
                 "event4",