+++ src/test_objects/mod.rs
@@ -1,3 +1,5 @@
+use tokio::task;
+
 use super::{EventProcessor, RealtimeStatus};
 use std::time::{Duration, Instant};